backend/requirements.txt
torch>=2.3.0
transformers>=4.38.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
Pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.4
matplotlib>=3.8.3
scikit-learn>=1.4.0
